requirements.txt
# Core dependencies
fastapi>=0.105.0
uvicorn>=0.25.0
pydantic>=2.5.2
python-dotenv>=1.0.0
aiohttp>=3.9.1
httpx>=0.25.2

# LLM providers
openai>=1.3.7
boto3>=1.33.6  # For AWS Bedrock
anthropic>=0.7.4  # For Anthropic Claude models
huggingface-hub>=0.19.4

# Web search providers
google-api-python-client>=2.108.0
serpapi>=0.1.0

# PDF processing
PyMuPDF>=1.23.7  # Also known as fitz
python-pptx>=0.6.22
Pillow>=10.1.0
tabula-py>=2.8.2  # For table extraction from PDFs
pytesseract>=0.3.10  # For OCR

# Data analysis
pandas>=2.1.3
numpy>=1.26.2
matplotlib>=3.8.2
seaborn>=0.13.0

# Vector databases
pinecone-client>=2.2.4
chromadb>=0.4.18
sentence-transformers>=2.2.2  # For embeddings

# Storage and serialization
boto3>=1.33.6  # For S3
sqlalchemy>=2.0.23  # For SQL database access

# UI
streamlit>=1.29.0
plotly>=5.18.0

# Security
python-jose>=3.3.0  # For JWT
passlib>=1.7.4  # For password hashing
bcrypt>=4.0.1  # For password hashing

# Monitoring and logging
prometheus-client>=0.17.1
python-json-logger>=2.0.7

# Code execution (sandboxed)
docker>=6.1.3  # For containerized code execution

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Development tools
black>=23.11.0
isort>=5.12.0
mypy>=1.7.1
flake8>=6.1.0

# Deployment
gunicorn>=21.2.0